Refactor IRForReleases test methods that into helper class
Since both test classes use the same methods it should be held
in a helper class.
